Remove ALL images whose left edge extends beyond the left edge of its parent.
This accounts for the possibility of multiple images extending beyond the limit in a single page load.  This originally occurred as a result of a bug, which has also been fixed in this commit, where the server was reading past the end of the list of gallery images and trying to send a non-existing image.